Move spec test ORM models out of MockBackend
This means that the schema and the models can live together, and are
isolated from the legacy backend definition.
